// src/app/login/interfaces/register.interface.ts
export interface Register {
  password:            string;
  username:            string;
  first_name:          string;
  last_name:           string;
  tipo:                string;
  nif:                 string;
  email_institucional: string;
  email_personal:      string;
  telefono:            string;
  curso:               string;
  titulacion:          number;
}